feat(ingest): M4 — analyze_image MCP tool + docs
Full pipeline callable live with structured report + inline grid
overlay (+ $ATARI2600_INGEST_PATH file). []uint8 in tool outputs
marshals as base64 and fails the generated schema — byte sequences
are []int now. Docs/README/CLAUDE.md routing updated; serverInfo
1.12.0. Smoke-tested initialize + analyze_image on the Pizza Boy shot
(29 bands, 16 sprites).

+++ internal/ingest/emit.go
@@ -60,7 +60,7 @@ func DASMSprites(sprites []Sprite) string {
 		if len(s.GRP) > 0 {
 			out += fmt.Sprintf("Spr%dGfx:\n", i)
 			for r, b := range s.GRP {
-				out += fmt.Sprintf("        byte %%%08b ; color $%02X\n", b, s.Colors[r])
+				out += fmt.Sprintf("        byte %%%08b ; color $%02X\n", uint8(b), uint8(s.Colors[r]))
 			}
 		}
 	}
